+++ Framework/Common/GraphicsManager.hpp
@@ -33,19 +33,19 @@ namespace My {
 
         virtual bool CheckCapability(RHICapability cap);
 
-        virtual void UseShaderProgram(const intptr_t shaderProgram);
+        virtual void UseShaderProgram(const int32_t shaderProgram);
         virtual void SetPerFrameConstants(const DrawFrameContext& context);
         virtual void SetPerBatchConstants(const std::vector<std::shared_ptr<DrawBatchContext>>& batches);
         virtual void SetLightInfo(const LightInfo& lightInfo);
 
         virtual void DrawBatch(const std::vector<std::shared_ptr<DrawBatchContext>>& batches);
 
-        virtual intptr_t GenerateCubeShadowMapArray(const uint32_t width, const uint32_t height, const uint32_t count);
-        virtual intptr_t GenerateShadowMapArray(const uint32_t width, const uint32_t height, const uint32_t count);
-        virtual void BeginShadowMap(const Light& light, const intptr_t shadowmap, const uint32_t width, const uint32_t height, const uint32_t layer_index);
-        virtual void EndShadowMap(const intptr_t shadowmap, const uint32_t layer_index);
+        virtual int32_t GenerateCubeShadowMapArray(const uint32_t width, const uint32_t height, const uint32_t count);
+        virtual int32_t GenerateShadowMapArray(const uint32_t width, const uint32_t height, const uint32_t count);
+        virtual void BeginShadowMap(const Light& light, const int32_t shadowmap, const uint32_t width, const uint32_t height, const uint32_t layer_index);
+        virtual void EndShadowMap(const int32_t shadowmap, const uint32_t layer_index);
         virtual void SetShadowMaps(const Frame& frame);
-        virtual void DestroyShadowMap(intptr_t& shadowmap);
+        virtual void DestroyShadowMap(int32_t& shadowmap);
 
         // skybox
         virtual void SetSkyBox(const DrawFrameContext& context);
@@ -55,14 +55,14 @@ namespace My {
         virtual void SetTerrain(const DrawFrameContext& context);
         virtual void DrawTerrain();
 
-        virtual intptr_t GenerateTexture(const char* id, const uint32_t width, const uint32_t height);
-        virtual void BeginRenderToTexture(intptr_t& context, const intptr_t texture, const uint32_t width, const uint32_t height);
-        virtual void EndRenderToTexture(intptr_t& context);
+        virtual int32_t GenerateTexture(const char* id, const uint32_t width, const uint32_t height);
+        virtual void BeginRenderToTexture(int32_t& context, const int32_t texture, const uint32_t width, const uint32_t height);
+        virtual void EndRenderToTexture(int32_t& context);
 
-        virtual intptr_t GenerateAndBindTextureForWrite(const char* id, const uint32_t width, const uint32_t height);
+        virtual int32_t GenerateAndBindTextureForWrite(const char* id, const uint32_t width, const uint32_t height);
         virtual void Dispatch(const uint32_t width, const uint32_t height, const uint32_t depth);
 
-        virtual intptr_t GetTexture(const char* id);
+        virtual int32_t GetTexture(const char* id);
 
         virtual void DrawFullScreenQuad();
 
@@ -76,10 +76,10 @@ namespace My {
         virtual void DrawTriangle(const PointList& vertices, const Vector3f &color);
         virtual void DrawTriangle(const PointList& vertices, const Matrix4X4f& trans, const Vector3f &color);
         virtual void DrawTriangleStrip(const PointList& vertices, const Vector3f &color);
-        virtual void DrawTextureOverlay(const intptr_t texture, float vp_left, float vp_top, float vp_width, float vp_height);
-        virtual void DrawTextureArrayOverlay(const intptr_t texture, uint32_t layer_index, float vp_left, float vp_top, float vp_width, float vp_height);
-        virtual void DrawCubeMapOverlay(const intptr_t cubemap, float vp_left, float vp_top, float vp_width, float vp_height, float level = 0.0f);
-        virtual void DrawCubeMapArrayOverlay(const intptr_t cubemap, uint32_t layer_index, float vp_left, float vp_top, float vp_width, float vp_height, float level = 0.0f);
+        virtual void DrawTextureOverlay(const int32_t texture, float vp_left, float vp_top, float vp_width, float vp_height);
+        virtual void DrawTextureArrayOverlay(const int32_t texture, uint32_t layer_index, float vp_left, float vp_top, float vp_width, float vp_height);
+        virtual void DrawCubeMapOverlay(const int32_t cubemap, float vp_left, float vp_top, float vp_width, float vp_height, float level = 0.0f);
+        virtual void DrawCubeMapArrayOverlay(const int32_t cubemap, uint32_t layer_index, float vp_left, float vp_top, float vp_width, float vp_height, float level = 0.0f);
         virtual void ClearDebugBuffers();
 
         void DrawEdgeList(const EdgeList& edges, const Vector3f& color);
